server/requirements.txt
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
